+++ src/extensions/taction/tactionext.cpp
@@ -154,6 +154,7 @@ bool TActionClassExtension::Execute(TActionClass& taction, HouseClass* house, Ob
         DISPATCH(TEXT_TRIGGER);
         DISPATCH(DESTROY_TRIGGER);
         DISPATCH(ENABLE_TRIGGER);
+        DISPATCH(DESTROY_TAG);
         DISPATCH(PLAY_SOUND_RANDOM);
 
         /**
@@ -240,6 +241,7 @@ bool TActionClassExtension::Is_Vinifera_TAction(TActionType type)
     case TACTION_TEXT_TRIGGER:
     case TACTION_DESTROY_TRIGGER:
     case TACTION_ENABLE_TRIGGER:
+    case TACTION_DESTROY_TAG:
     case TACTION_PLAY_SOUND_RANDOM:
         return true;
 
@@ -458,6 +460,26 @@ bool TActionClassExtension::Do_ENABLE_TRIGGER(TActionClass& taction, HouseClass*
                 if (really_enable) {
                     Triggers[index]->Enable();
                 }
+            }
+        }
+    }
+    return true;
+}
+
+
+/**
+ *  Fixes a bug where caching Tags.Count() before the loop
+ *  could cause out-of-bounds access after deletions
+ *
+ *  @author: ZivDero
+ */
+bool TActionClassExtension::Do_DESTROY_TAG(TActionClass& taction, HouseClass* house, ObjectClass* object, TriggerClass* trig, const Cell& cell)
+{
+    if (taction.Tag != nullptr) {
+        for (int index = 0; index < Tags.Count(); index++) {
+            if (Tags[index]->Class == taction.Tag) {
+                delete Tags[index];
+                index--;
             }
         }
     }