+++ docs/Whats-New.md
@@ -87,6 +87,7 @@ Vanilla fixes:
 - Fix a bug where the player couldn't repair aircraft on allied repair bays under some circumstances (by ZivDero)
 - Fix a buffer overflow crash in `MapClass::Place_Down` (by Rampastring)
 - Fix a bug where the "Building exists" event would fire when you queued a building on the sidebar (by ZivDero)
+- Fix a bug where using the "Destroy Tag" trigger action could lead to trying to free invalid memory (by ZivDero)
 
 :::